feat: reorganized structure
