+++ s01/c01/test_c01.py
@@ -1,0 +1,11 @@
+#
+#   01 - Convert hex to base64
+#
+
+from c01 import c01
+
+def test_c01() -> None:
+    input = b"49276d206b696c6c696e6720796f757220627261696e206c696b65206120706f69736f6e6f7573206d757368726f6f6d"
+    result = b"SSdtIGtpbGxpbmcgeW91ciBicmFpbiBsaWtlIGEgcG9pc29ub3VzIG11c2hyb29t"
+    
+    assert c01(input) == result
