+++ s01/c05/helper_c05.py
@@ -1,0 +1,9 @@
+#
+#   02 - Fixed XOR
+#
+
+def fixed_xor(input1: bytes, input2: bytes) -> bytes:
+    if len(input1) == len(input2):
+        return bytes(a ^ b for (a, b) in zip(input1, input2))
+    else:
+        raise ValueError
