+++ s01/c06/6.txt
