+++ s02/c10/10.txt
