+++ s02/c12/test_c12.py
@@ -1,0 +1,10 @@
+#
+#   12 - Byte-at-a-time ECB decryption (Simple)
+#
+
+from c12 import c12
+
+def test_c12() -> None:
+    result = b"Um9sbGluJyBpbiBteSA1LjAKV2l0aCBteSByYWctdG9wIGRvd24gc28gbXkgaGFpciBjYW4gYmxvdwpUaGUgZ2lybGllcyBvbiBzdGFuZGJ5IHdhdmluZyBqdXN0IHRvIHNheSBoaQpEaWQgeW91IHN0b3A/IE5vLCBJIGp1c3QgZHJvdmUgYnkK"
+    
+    assert c12() == result
