+++ s02/c14/test_c14.py
@@ -1,0 +1,10 @@
+#
+#   14 - Byte-at-a-time ECB decryption (Harder)
+#
+
+from c14 import c14
+
+def test_c14() -> None:
+    result = b"Um9sbGluJyBpbiBteSA1LjAKV2l0aCBteSByYWctdG9wIGRvd24gc28gbXkgaGFpciBjYW4gYmxvdwpUaGUgZ2lybGllcyBvbiBzdGFuZGJ5IHdhdmluZyBqdXN0IHRvIHNheSBoaQpEaWQgeW91IHN0b3A/IE5vLCBJIGp1c3QgZHJvdmUgYnkK"
+    
+    assert c14() == result
